% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{acf_cpp}
\alias{acf_cpp}
\title{Fast ACF computation (C++ implementation)}
\usage{
acf_cpp(x, max_lag)
}
\arguments{
\item{x}{Numeric vector}

\item{max_lag}{Maximum lag}
}
\value{
NumericVector of ACF values
}
\description{
Efficient computation of autocorrelation function.
}
