\name{chaoticds-all}
\docType{package}
\title{Documentation for chaoticds Package Functions}
\description{This file provides minimal documentation for exported functions in the \pkg{chaoticds} package.}
\details{
The functions provide tools for simulating chaotic maps, estimating extremal indices, and related utilities. For usage examples see the package vignette or README.
}
\author{Diogo Ribeiro}
\keyword{internal}
\alias{simulate_logistic_map}
\alias{simulate_henon_map}
\alias{simulate_tent_map}
\alias{simulate_lozi_map}
\alias{logistic_bifurcation}
\alias{estimate_lyapunov_exponent}
\alias{block_maxima}
\alias{fit_gev}
\alias{exceedances}
\alias{fit_gpd}
\alias{mean_residual_life}
\alias{mrl_plot}
\alias{threshold_diagnostics}
\alias{hill_estimates}
\alias{hill_plot}
\alias{bootstrap_extremal_index}
\alias{cluster_sizes}
\alias{cluster_summary}
\alias{cluster_histogram}
\alias{acf_decay}
\alias{mixing_coefficients}
\alias{d_check}
\alias{threshold_exceedances}
\alias{cluster_exceedances}
\alias{extremal_index_runs}
\alias{extremal_index_intervals}
\alias{hitting_times}
\alias{plot_hts}
\alias{run_demo}
\alias{launch_explorer}
\alias{simulate_logistic_map_fast}
\alias{extremal_index_runs_fast}
\alias{block_maxima_fast}
\alias{benchmark_implementations}
\alias{performance_analysis}
\alias{recurrence_plot}
\alias{recurrence_analysis}
\alias{simulate_cat_map}
\alias{clean_extreme_data}
\alias{empirical_quantile}
\alias{compute_autocorrelation}
\alias{with_logging}

