% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{block_maxima_cpp}
\alias{block_maxima_cpp}
\title{Fast block maxima computation (C++ implementation)}
\usage{
block_maxima_cpp(x, block_size)
}
\arguments{
\item{x}{Numeric vector of data}

\item{block_size}{Size of each block}
}
\value{
NumericVector of block maxima
}
\description{
Efficient C++ implementation for computing block maxima.
}
