% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecdf_cpp}
\alias{ecdf_cpp}
\title{Fast empirical CDF (C++ implementation)}
\usage{
ecdf_cpp(x, eval_points)
}
\arguments{
\item{x}{Numeric vector (data)}

\item{eval_points}{Numeric vector (points to evaluate CDF)}
}
\value{
NumericVector of CDF values
}
\description{
Compute empirical cumulative distribution function.
}
