% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empirical_quantile}
\alias{empirical_quantile}
\title{Empirical quantile}
\usage{
empirical_quantile(x, prob)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{prob}{Numeric probability in `[0, 1]` specifying the quantile.}
}
\value{
Numeric value giving the empirical quantile.
}
\description{
Computes the empirical quantile of a numeric vector using `type = 8`
which is recommended for statistical applications.
}
\examples{
empirical_quantile(rnorm(100), 0.95)
}
