% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppexcesses.R
\name{exceedance_indices}
\alias{exceedance_indices}
\title{Identify Exceedance Indices}
\usage{
exceedance_indices(X, u)
}
\arguments{
\item{X}{Numeric vector. Time series.}

\item{u}{Numeric scalar. Threshold.}
}
\value{
Integer vector of time indices where X > u.
}
\description{
Identify Exceedance Indices
}
