% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks-over-threshold.R
\name{exceedances}
\alias{exceedances}
\title{Identify exceedances above a threshold}
\usage{
exceedances(x, threshold)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{threshold}{Numeric threshold value.}
}
\value{
Numeric vector of exceedances (values above `threshold`). If no
  values exceed `threshold` an empty vector is returned.
}
\description{
Identify exceedances above a threshold
}
\examples{
exceedances(rnorm(100), 1.5)
}
\seealso{
[fit_gpd()] for fitting GPD to exceedances,
  [threshold_exceedances()] for indices of exceedances,
  [block_maxima()] for alternative block maxima approach
}
