% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exceedances_cpp}
\alias{exceedances_cpp}
\title{Fast exceedances extraction (C++ implementation)}
\usage{
exceedances_cpp(x, threshold)
}
\arguments{
\item{x}{Numeric vector}

\item{threshold}{Threshold value}
}
\value{
NumericVector of exceedance values
}
\description{
Extract values that exceed threshold.
}
