% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremal-index.R
\name{extremal_index}
\alias{extremal_index}
\title{Extremal index and hitting time utilities}
\description{
Implements estimation of the extremal index and empirical return time
statistics for a univariate stationary time series.
}
