% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extremal_index_intervals_cpp}
\alias{extremal_index_intervals_cpp}
\title{Fast extremal index intervals estimator (C++ implementation)}
\usage{
extremal_index_intervals_cpp(x, threshold)
}
\arguments{
\item{x}{Numeric vector}

\item{threshold}{Numeric threshold}
}
\value{
Numeric extremal index estimate
}
\description{
Ferro-Segers intervals estimator for extremal index.
}
