% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast-functions.R
\name{fast-functions}
\alias{fast-functions}
\title{Fast simulation and analysis functions}
\description{
Wrapper functions that automatically choose between R and C++ implementations
based on data size and availability of compiled code.
}
