% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppexcesses.R
\name{fit_compound_poisson}
\alias{fit_compound_poisson}
\title{Fit Compound Poisson Process}
\usage{
fit_compound_poisson(mpp, period)
}
\arguments{
\item{mpp}{Data.frame with time and mark columns.}

\item{period}{Length of observation window.}
}
\value{
List with lambda estimate and mark distribution fit.
}
\description{
Fit Compound Poisson Process
}
