% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold-selection.R
\name{hill_plot}
\alias{hill_plot}
\title{Plot Hill estimates}
\usage{
hill_plot(hill_df)
}
\arguments{
\item{hill_df}{Data frame as returned by [hill_estimates()].}
}
\value{
ggplot object visualizing the Hill plot. Requires the **ggplot2**
  package.
}
\description{
Plot Hill estimates
}
\examples{
df <- hill_estimates(rexp(1000), 1:50)
hill_plot(df)
}
