% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate-extremes.R
\name{lower_tail_dependence}
\alias{lower_tail_dependence}
\title{Lower tail dependence}
\usage{
lower_tail_dependence(x, y, ux, uy)
}
\arguments{
\item{x}{[numeric] Vector of observations.}

\item{y}{[numeric] Vector of the same length as `x`.}

\item{ux}{[numeric] Threshold for `x`.}

\item{uy}{[numeric] Threshold for `y`.}
}
\value{
[numeric] Lower tail dependence coefficient.
}
\description{
Convenience wrapper for [tail_dependence_asymmetric()] computing
lower tail dependence with possibly different quantile levels.
}
