% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moving_average_cpp}
\alias{moving_average_cpp}
\title{Fast moving average (C++ implementation)}
\usage{
moving_average_cpp(x, window_size)
}
\arguments{
\item{x}{Numeric vector}

\item{window_size}{Integer window size}
}
\value{
NumericVector of moving averages
}
\description{
Compute moving average for time series.
}
