% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks-over-threshold.R
\name{mrl_plot}
\alias{mrl_plot}
\title{Plot Mean Residual Life (MRL)}
\usage{
mrl_plot(mrl_df)
}
\arguments{
\item{mrl_df}{Data frame as returned by [mean_residual_life()].}
}
\value{
ggplot object visualizing the MRL curve. Requires the **ggplot2**
  package.
}
\description{
Plot Mean Residual Life (MRL)
}
\examples{
df <- mean_residual_life(rnorm(1000), seq(0, 2, 0.2))
mrl_plot(df)
}
