% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate-extremes.R
\name{multivariate_extremes}
\alias{multivariate_extremes}
\title{Multivariate Extreme Value Utilities}
\description{
Collection of functions for multivariate extremal analysis and
tail dependence diagnostics.
}
