% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimized-wrappers.R
\name{print.bootstrap_ei}
\alias{print.bootstrap_ei}
\title{Print method for bootstrap_ei}
\usage{
\method{print}{bootstrap_ei}(x, ...)
}
\arguments{
\item{x}{Object of class bootstrap_ei}

\item{...}{Additional arguments}
}
\description{
Print method for bootstrap_ei
}
