% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{print.chaotic_model}
\alias{print.chaotic_model}
\title{Print chaoticds model summary header}
\usage{
\method{print}{chaotic_model}(x, ...)
}
\arguments{
\item{x}{Fitted model object returned by [fit_gev()] or [fit_gpd()].}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print chaoticds model summary header
}
