% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile-likelihood.R
\name{print.profile_likelihood}
\alias{print.profile_likelihood}
\title{Print method for profile-likelihood objects}
\usage{
\method{print}{profile_likelihood}(x, ...)
}
\arguments{
\item{x}{A `profile_likelihood` object.}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print method for profile-likelihood objects
}
