% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-api.R
\name{print.summary.chaotic_model}
\alias{print.summary.chaotic_model}
\title{Print method for [summary.chaotic_model()]}
\usage{
\method{print}{summary.chaotic_model}(x, ...)
}
\arguments{
\item{x}{Output of [summary.chaotic_model()].}

\item{...}{Unused.}
}
\value{
Invisibly returns `x`.
}
\description{
Print method for [summary.chaotic_model()]
}
