% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_cpp}
\alias{quantile_cpp}
\title{Fast quantile computation (C++ implementation)}
\usage{
quantile_cpp(x, prob)
}
\arguments{
\item{x}{Numeric vector of data}

\item{prob}{Probability (0 to 1)}
}
\value{
Numeric scalar quantile value
}
\description{
Simple quantile computation for threshold selection.
}
