% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{return_level_empirical_cpp}
\alias{return_level_empirical_cpp}
\title{Fast return level estimation (C++ implementation)}
\usage{
return_level_empirical_cpp(block_maxima, return_period)
}
\arguments{
\item{block_maxima}{Numeric vector of block maxima}

\item{return_period}{Integer return period}
}
\value{
Numeric scalar return level estimate
}
\description{
Estimate return levels from block maxima.
}
