% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppexcesses.R
\name{select_threshold}
\alias{select_threshold}
\title{Select High Threshold via Quantile}
\usage{
select_threshold(X, prob = 0.99)
}
\arguments{
\item{X}{Numeric vector. Time series.}

\item{prob}{Numeric in (0,1). Exceedance probability level.}
}
\value{
Numeric scalar. Threshold u.
}
\description{
Select High Threshold via Quantile
}
