% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppexcesses.R
\name{simulate_orbit}
\alias{simulate_orbit}
\title{Simulate Orbit of a One-Dimensional Map}
\usage{
simulate_orbit(map_fn, init, n_iter)
}
\arguments{
\item{map_fn}{Function(x) returning numeric scalar. The dynamical map f: X->[0,1].}

\item{init}{Numeric scalar in domain of f. Initial condition.}

\item{n_iter}{Integer > 0. Number of iterations.}
}
\value{
Numeric vector of length n_iter+1 (includes init). Orbit values.
}
\description{
Simulate Orbit of a One-Dimensional Map
}
\examples{
orb <- simulate_orbit(function(x) 4*x*(1-x), init=0.1, n_iter=1e5)
}
