% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppexcesses.R
\name{marked_point_process}
\alias{marked_point_process}
\title{Build Marked Point Process}
\usage{
marked_point_process(
  X,
  u,
  run_length = 0L,
  type = c("REPP", "EOT", "POT", "AOT")
)
}
\arguments{
\item{X}{Numeric vector. Time series.}

\item{u}{Numeric scalar. Threshold.}

\item{run_length}{Integer. Cluster run-length p.}

\item{type}{Character: one of "REPP","EOT","POT","AOT".}
}
\value{
Data.frame with columns time (start of cluster) and mark.
}
\description{
Build Marked Point Process
}
