% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_logistic_map_cpp}
\alias{simulate_logistic_map_cpp}
\title{Fast logistic map simulation (C++ implementation)}
\usage{
simulate_logistic_map_cpp(n, r, x0)
}
\arguments{
\item{n}{Number of iterations}

\item{r}{Parameter r of the logistic map}

\item{x0}{Initial value}
}
\value{
Numeric vector of the time series
}
\description{
Efficient C++ implementation of the logistic map for improved performance
with large datasets.
}
