% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_lozi_map_cpp}
\alias{simulate_lozi_map_cpp}
\title{Fast Lozi map simulation (C++ implementation)}
\usage{
simulate_lozi_map_cpp(n, a = 1.7, b = 0.5, x0 = 0, y0 = 0)
}
\arguments{
\item{n}{Number of iterations}

\item{a}{Parameter a}

\item{b}{Parameter b}

\item{x0}{Initial x value}

\item{y0}{Initial y value}
}
\value{
DataFrame with x and y columns
}
\description{
Efficient C++ implementation of the two-dimensional Lozi map.
}
