% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_tent_map_cpp}
\alias{simulate_tent_map_cpp}
\title{Fast tent map simulation (C++ implementation)}
\usage{
simulate_tent_map_cpp(n, r, x0)
}
\arguments{
\item{n}{Number of iterations}

\item{r}{Parameter r}

\item{x0}{Initial value}
}
\value{
Numeric vector
}
\description{
Efficient C++ implementation of the tent map.
}
